fix: additional int64 conversions for Phase 1
- NDArray constructors: int size -> long size
- NDArray.GetAtIndex/SetAtIndex: int index -> long index
- UnmanagedStorage.GetAtIndex/SetAtIndex: int index -> long index
- ValueCoordinatesIncrementor.Next(): int[] -> long[]
- DefaultEngine.MoveAxis: int[] -> long[]

Build still failing - cascading changes needed in:
- All incrementors (NDCoordinatesIncrementor, NDOffsetIncrementor, etc.)
- NDIterator and all cast files
- UnmanagedStorage.Cloning
- np.random.shuffle, np.random.choice

Related to #584

+++ src/NumSharp.Core/LinearAlgebra/np.linalg.norm.cs
@@ -78,7 +78,7 @@ namespace NumSharp
                 else if (axis_obj is int)
                     axis = new int[] {(int)axis_obj};
                 else if (axis_obj is int[])
-                    axis = (int[])axis_obj;
+                    axis = (long[])axis_obj;
                 else
                     throw new ArgumentException($"Invalid axis type: {axis_obj}");
 //                if (axis.Length == 1)
